refactor: extract download proxy route
